refactor(tests): Centralize test utility functions
Refactors the test suite to centralize duplicated test utility functions.

Creates a new `tests/test-utils` directory to house common test helpers, including data factories and bluetooth mocks. Updates all relevant test files to import from this new centralized module.

This improves maintainability, reduces code duplication, and enhances consistency across the test suite.
