+++ tests/test-utils/bluetooth-test-utils.ts
